requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Progetti Github\RetniNet\index.py: 26
bs4 == 0.0.1

# D:\Progetti Github\RetniNet\index.py: 23
jmespath == 1.0.0

# D:\Progetti Github\RetniNet\index.py: 1
psutil == 5.9.1

# D:\Progetti Github\RetniNet\index.py: 3
pyfiglet == 0.8.post1

# D:\Progetti Github\RetniNet\index.py: 11
python_ffmpeg == 1.0.14

# D:\Progetti Github\RetniNet\index.py: 14
python_weather == 0.3.7

# D:\Progetti Github\RetniNet\index.py: 25
pytube == 12.1.0

# D:\Progetti Github\RetniNet\index.py: 13,16
requests == 2.25.1

# D:\Progetti Github\RetniNet\index.py: 24
telebot == 0.0.4

# D:\Progetti Github\RetniNet\index.py: 20
tqdm == 4.64.0

# D:\Progetti Github\RetniNet\index.py: 18
yt_dlp == 2022.5.18
